Refactor include paths
