+++ clients/c_cpp/csal/csal_attrib.cpp
@@ -1,6 +1,6 @@
 
 #include "sal.h"
-#include "csal.h"
+#include "csal/csal.h"
 //#include "sal.h"
