+++ clients/c_cpp/csal/include/csal/detail/csal_attrib.h
